py: improve numeric string conversion

+++ transpiler/x/py/transpiler.go
@@ -154,7 +154,7 @@ import builtins
 def _str(v):
     if isinstance(v, float):
         if abs(v - builtins.round(v)) < 1e-9:
-            return builtins.str(float(builtins.round(v)))
+            return builtins.str(int(builtins.round(v)))
         return builtins.format(v, ".15g")
     return builtins.str(v)
 `
